Made a Tutorial Conversation with Father. (TEMP) Also made some stream lining for creation of more dialogue!
There's a template for the dialogue trees now. Just go into the explorer and copy paste the template wherever you want. Then go in and change the dialogue in the animator.

I also made a prefab of the "Dialogue Object" which can be slapped onto any npc and should get the basis for dialogue setup. (can always be improved ofc)